Actually split API routes by domain and remove unnecessary shims
