+++ src/barsukas/routes/api/lemma_routes.py
@@ -57,319 +57,6 @@ from workqueue.task_queue import TaskStatus, TaskType, enqueue_task
 from barsukas.routes.api import bp
 
 
-@bp.route("/v1")
-def api_info() -> ResponseReturnValue:
-    """
-    Get information about the API and available endpoints.
-
-    Returns:
-        - version: API version
-        - endpoints: List of available endpoints with descriptions
-    """
-    endpoints = [
-        {
-            "path": "/api/v1/search",
-            "method": "GET",
-            "description": "Search for lemmas by keyword",
-            "parameters": [
-                {
-                    "name": "q",
-                    "type": "query",
-                    "required": True,
-                    "description": "Search query (searches lemma text, definition, disambiguation, translations)",
-                },
-                {
-                    "name": "pos_type",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter by part of speech (e.g., 'noun', 'verb')",
-                },
-                {
-                    "name": "difficulty",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter by difficulty level (1-30, '-1' for excluded, 'null' for not set)",
-                },
-                {
-                    "name": "missing_translation",
-                    "type": "query",
-                    "required": False,
-                    "description": "Only return lemmas missing this target language translation",
-                },
-                {
-                    "name": "limit",
-                    "type": "query",
-                    "required": False,
-                    "description": "Max results to return (default: 20, max: 100)",
-                },
-                {
-                    "name": "offset",
-                    "type": "query",
-                    "required": False,
-                    "description": "Number of results to skip for pagination (default: 0)",
-                },
-            ],
-        },
-        {
-            "path": "/api/v1/lemmas/by-difficulty",
-            "method": "GET",
-            "description": "List lemmas by difficulty level without requiring a text query",
-            "parameters": [
-                {
-                    "name": "difficulty",
-                    "type": "query",
-                    "required": True,
-                    "description": "Difficulty level (1-30, '-1' for excluded, 'null' for not set)",
-                },
-                {
-                    "name": "pos_type",
-                    "type": "query",
-                    "required": False,
-                    "description": "Optional part-of-speech filter",
-                },
-                {
-                    "name": "limit",
-                    "type": "query",
-                    "required": False,
-                    "description": "Max results to return (default: 200, max: 1000)",
-                },
-                {
-                    "name": "offset",
-                    "type": "query",
-                    "required": False,
-                    "description": "Number of results to skip for pagination (default: 0)",
-                },
-            ],
-        },
-        {
-            "path": "/api/v1/lemmas/translations",
-            "method": "GET",
-            "description": "Fetch translations for multiple lemma GUIDs in one request",
-            "parameters": [
-                {
-                    "name": "guids",
-                    "type": "query",
-                    "required": True,
-                    "description": "Comma-separated GUID list",
-                },
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Optional language filter",
-                },
-            ],
-        },
-        {
-            "path": "/api/v1/models",
-            "method": "GET",
-            "description": "List LLM models registered in the benchmarks database (requires postgres)",
-            "parameters": [
-                {
-                    "name": "q",
-                    "type": "query",
-                    "required": False,
-                    "description": "Case-insensitive substring search across codename, displayname, model_path, lmstudio_model_name",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/metadata/words",
-            "method": "GET",
-            "description": "Get per-language lemma counts and metadata coverage (audio, derivative forms, subtypes)",
-            "parameters": [
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter to a specific language code (e.g., 'en', 'zh', 'lt')",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/metadata/pos-subtypes",
-            "method": "GET",
-            "description": "List available POS subtypes, optionally filtered by POS type",
-            "parameters": [
-                {
-                    "name": "pos_type",
-                    "type": "query",
-                    "required": False,
-                    "description": "Optional part-of-speech filter (e.g., noun, verb)",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/metadata/levels/by-pos",
-            "method": "GET",
-            "description": "Get word difficulty-level distribution for a POS type + subtype",
-            "parameters": [
-                {
-                    "name": "pos_type",
-                    "type": "query",
-                    "required": True,
-                    "description": "Part-of-speech type (e.g., noun, verb)",
-                },
-                {
-                    "name": "pos_subtype",
-                    "type": "query",
-                    "required": True,
-                    "description": "Part-of-speech subtype (e.g., physical_object)",
-                },
-            ],
-        },
-        {
-            "path": "/api/v1/metadata/sentences",
-            "method": "GET",
-            "description": "Get per-language sentence counts and metadata coverage (audio, verification, pattern types)",
-            "parameters": [
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter to a specific language code (e.g., 'en', 'zh', 'lt')",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/sentences/add",
-            "method": "POST",
-            "description": "Add up to 15 sentences to the database, deduplicating by text. Returns IDs for use with /api/llm/sentences/decompose.",
-            "parameters": [
-                {
-                    "name": "sentences",
-                    "type": "body",
-                    "required": True,
-                    "description": "List of sentence strings (max 15)",
-                },
-                {
-                    "name": "source",
-                    "type": "body",
-                    "required": True,
-                    "description": "Source identifier stored as source_filename on each created Sentence row",
-                },
-                {
-                    "name": "language",
-                    "type": "body",
-                    "required": False,
-                    "description": "Source language code of the input sentences (default: en)",
-                },
-            ],
-        },
-        {
-            "path": "/api/v1/lemma/<guid>",
-            "method": "GET",
-            "description": "Get basic information about a lemma",
-            "parameters": [],
-        },
-        {
-            "path": "/api/v1/lemma/<guid>/translations",
-            "method": "GET",
-            "description": "Get translations of a lemma in various languages",
-            "parameters": [
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter to specific language code (e.g., 'zh', 'fr', 'lt')",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/lemma/<guid>/forms",
-            "method": "GET",
-            "description": "Get derivative/declined forms of a lemma",
-            "parameters": [
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter to specific language code (e.g., 'zh', 'fr', 'lt')",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/lemma/<guid>/grammar",
-            "method": "GET",
-            "description": "Get grammar facts about a lemma",
-            "parameters": [
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter to specific language code (e.g., 'zh', 'fr', 'lt')",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/lemma/<guid>/pronunciations",
-            "method": "GET",
-            "description": "Get pronunciations (IPA and phonetic) for base forms of a lemma",
-            "parameters": [
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter to specific language code (e.g., 'en', 'lt', 'fr')",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/lemma/<guid>/audio",
-            "method": "GET",
-            "description": "Get lemma-level and form-level audio availability by language",
-            "parameters": [
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter to specific language code (e.g., 'en', 'lt', 'fr')",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/audio/voices",
-            "method": "GET",
-            "description": "List available TTS voices by backend and language",
-            "parameters": [
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter to specific language code (e.g., 'bs')",
-                }
-            ],
-        },
-        {
-            "path": "/api/v1/lemma/<guid>/wordfreq",
-            "method": "GET",
-            "description": "Get per-language wordfreq corpus rollups and best ranks for a lemma",
-            "parameters": [],
-        },
-        {
-            "path": "/api/v1/lemma/<guid>/sentences",
-            "method": "GET",
-            "description": "Get example sentences that use this lemma",
-            "parameters": [
-                {
-                    "name": "language",
-                    "type": "query",
-                    "required": False,
-                    "description": "Filter sentence translations to specific language code",
-                }
-            ],
-        },
-    ]
-
-    return jsonify(
-        {
-            "version": "1.0",
-            "description": "Read-only API for accessing lemma information",
-            "endpoints": endpoints,
-        }
-    )
-
-
 @bp.route("/v1/search")
 @mirrored_facade("/api/v1/search", "GET")
 def search_lemmas() -> ResponseReturnValue:
@@ -1942,361 +1629,3 @@ def get_word_metadata() -> ResponseReturnValue:
         metadata["difficulty"] = exact_difficulty
 
     return _build_success_response(summary_data, metadata)
-
-
-@bp.route("/v1/metadata/pos-subtypes")
-@mirrored_facade("/api/v1/metadata/pos-subtypes", "GET")
-def list_pos_subtypes() -> ResponseReturnValue:
-    """List distinct POS subtypes, optionally filtered by POS type."""
-    pos_type = request.args.get("pos_type", "").strip()
-    subtype_query = g.db.query(Lemma.pos_subtype).filter(
-        Lemma.pos_subtype.isnot(None),
-        Lemma.pos_subtype != "",
-    )
-    if pos_type:
-        subtype_query = subtype_query.filter(Lemma.pos_type == pos_type)
-
-    subtypes = sorted({row[0] for row in subtype_query.distinct().all()})
-    metadata: Dict[str, Any] = {
-        "count": len(subtypes),
-    }
-    if pos_type:
-        metadata["pos_type"] = pos_type
-    return _build_success_response(subtypes, metadata)
-
-
-@bp.route("/v1/metadata/levels/by-pos")
-@mirrored_facade("/api/v1/metadata/levels/by-pos", "GET")
-def get_level_distribution_by_pos() -> ResponseReturnValue:
-    """Return difficulty distribution for lemmas in one POS type/subtype bucket."""
-    pos_type = request.args.get("pos_type", "").strip()
-    pos_subtype = request.args.get("pos_subtype", "").strip()
-    if not pos_type:
-        return _build_error_response("pos_type is required", 400)
-    if not pos_subtype:
-        return _build_error_response("pos_subtype is required", 400)
-
-    rows = (
-        g.db.query(Lemma.difficulty_level, func.count().label("count"))
-        .filter(Lemma.pos_type == pos_type, Lemma.pos_subtype == pos_subtype)
-        .group_by(Lemma.difficulty_level)
-        .order_by(Lemma.difficulty_level)
-        .all()
-    )
-    distribution = {
-        "null" if row.difficulty_level is None else str(row.difficulty_level): row.count
-        for row in rows
-    }
-    total_words = sum(row.count for row in rows)
-
-    return _build_success_response(
-        distribution,
-        {
-            "pos_type": pos_type,
-            "pos_subtype": pos_subtype,
-            "total_words": total_words,
-        },
-    )
-
-
-@bp.route("/v1/metadata/sentences")
-@mirrored_facade("/api/v1/metadata/sentences", "GET")
-def get_sentence_metadata() -> ResponseReturnValue:
-    """Get per-language metadata counts for sentence coverage and sentence-level audio.
-
-    Optional query params:
-      language=<code>       Filter to a single language.
-      max_difficulty=<int>  Only count sentences whose minimum_level is between
-                            1 and max_difficulty inclusive (excludes NULL / unset).
-    """
-    requested_language = request.args.get("language", "").strip().lower()
-    max_difficulty_param = request.args.get("max_difficulty", "").strip()
-    max_difficulty_sentences: Optional[int] = None
-    if max_difficulty_param:
-        try:
-            max_difficulty_sentences = int(max_difficulty_param)
-        except ValueError:
-            return _build_error_response("max_difficulty must be an integer", 400)
-
-    sentence_translation_languages = {
-        row[0]
-        for row in g.db.query(SentenceTranslation.language_code)
-        .filter(
-            SentenceTranslation.translation_text.isnot(None),
-            SentenceTranslation.translation_text != "",
-        )
-        .distinct()
-        .all()
-    }
-    sentence_audio_languages = {
-        row[0]
-        for row in g.db.query(AudioQualityReview.language_code)
-        .filter(AudioQualityReview.sentence_id.isnot(None))
-        .distinct()
-        .all()
-    }
-
-    available_languages = (
-        set(LANGUAGE_HIERARCHY) | sentence_translation_languages | sentence_audio_languages
-    )
-
-    if requested_language and requested_language not in available_languages:
-        return _build_error_response(
-            f"Language '{requested_language}' not found in metadata sources",
-            404,
-        )
-
-    ordered_languages = [lang for lang in LANGUAGE_HIERARCHY if lang in available_languages]
-    extra_languages = sorted(lang for lang in available_languages if lang not in LANGUAGE_HIERARCHY)
-    all_languages = ordered_languages + extra_languages
-
-    selected_languages = [requested_language] if requested_language else all_languages
-
-    summary_data: Dict[str, Any] = {}
-
-    for current_language in selected_languages:
-        sentence_filters = [
-            SentenceTranslation.language_code == current_language,
-            SentenceTranslation.translation_text.isnot(None),
-            SentenceTranslation.translation_text != "",
-        ]
-        if max_difficulty_sentences is not None:
-            sentence_filters.extend(
-                [
-                    Sentence.minimum_level.isnot(None),
-                    Sentence.minimum_level >= 1,
-                    Sentence.minimum_level <= max_difficulty_sentences,
-                ]
-            )
-        base_query = (
-            g.db.query(Sentence.id, Sentence.pattern_type, Sentence.verified)
-            .join(SentenceTranslation, SentenceTranslation.sentence_id == Sentence.id)
-            .filter(*sentence_filters)
-            .distinct()
-        )
-
-        base_subquery = base_query.subquery()
-
-        total_sentences = g.db.query(func.count()).select_from(base_subquery).scalar() or 0
-
-        pattern_rows = (
-            g.db.query(
-                case(
-                    (base_subquery.c.pattern_type.is_(None), "unspecified"),
-                    else_=base_subquery.c.pattern_type,
-                ).label("pattern_type"),
-                func.count().label("count"),
-            )
-            .group_by("pattern_type")
-            .order_by("pattern_type")
-            .all()
-        )
-        sentences_by_pattern = {row.pattern_type: row.count for row in pattern_rows}
-
-        audio_count = (
-            g.db.query(func.count(func.distinct(base_subquery.c.id)))
-            .select_from(base_subquery)
-            .join(
-                AudioQualityReview,
-                and_(
-                    AudioQualityReview.sentence_id == base_subquery.c.id,
-                    AudioQualityReview.language_code == current_language,
-                ),
-            )
-            .scalar()
-            or 0
-        )
-
-        verified_count = (
-            g.db.query(func.count())
-            .select_from(base_subquery)
-            .filter(base_subquery.c.verified.is_(True))
-            .scalar()
-            or 0
-        )
-
-        summary_data[current_language] = {
-            "total_sentences": total_sentences,
-            "sentences_by_pattern": sentences_by_pattern,
-            "audio": {
-                "with_audio": audio_count,
-                "without_audio": total_sentences - audio_count,
-            },
-            "verification": {
-                "verified": verified_count,
-                "unverified": total_sentences - verified_count,
-            },
-        }
-
-    metadata: Dict[str, Any] = {
-        "languages": selected_languages,
-        "count": len(selected_languages),
-        "notes": {
-            "sentence_definition": "A sentence with non-empty sentence_translations.translation_text in that language",
-            "audio_definition": "At least one sentence-level audio_quality_reviews row for the sentence and language (sentence_id set)",
-            "verification_definition": "Sentence.verified value among sentences counted in that language",
-        },
-    }
-
-    if requested_language:
-        metadata["requested_language"] = requested_language
-    if max_difficulty_sentences is not None:
-        metadata["max_difficulty"] = max_difficulty_sentences
-
-    return _build_success_response(summary_data, metadata)
-
-
-_ADD_SENTENCES_MAX = 15
-
-
-@bp.route("/v1/sentences/add", methods=["POST"])
-@mirrored_facade("/api/v1/sentences/add", "POST")
-def add_sentences() -> ResponseReturnValue:
-    """Add up to 15 sentences to the database, deduplicating by text.
-
-    Sentences already present (matched by text in the source language) are
-    returned with status ``already_exists`` so the caller can still obtain
-    their IDs. New sentences receive the caller-supplied source identifier
-    as ``source_filename``.
-
-    This endpoint makes no LLM calls. Use ``POST /api/llm/sentences/decompose``
-    afterward to queue translation + decomposition for the returned IDs.
-
-    Request body (JSON):
-        sentences (list[str], required): Up to 15 sentence strings.
-        source (str, required): Source identifier stored as source_filename.
-        language (str, optional): Source language code (default: "en").
-
-    Returns:
-        data: list of per-sentence results, each with:
-            - sentence_text: the input text
-            - sentence_id: database ID
-            - guid: sentence GUID (e.g. S_00123)
-            - status: "created" or "already_exists"
-        metadata: total, created, already_exists counts plus source and language.
-
-    Example:
-        POST /api/v1/sentences/add
-        {"sentences": ["The cat sits on the mat."], "source": "lesson_a1"}
-    """
-    payload = request.get_json(silent=True) or {}
-
-    sentences_raw = payload.get("sentences")
-    if not isinstance(sentences_raw, list) or not sentences_raw:
-        return _build_error_response("sentences must be a non-empty list")
-    if len(sentences_raw) > _ADD_SENTENCES_MAX:
-        return _build_error_response(f"sentences may contain at most {_ADD_SENTENCES_MAX} items")
-    sentences: List[str] = []
-    for item in sentences_raw:
-        if not isinstance(item, str) or not item.strip():
-            return _build_error_response("each sentence must be a non-empty string")
-        sentences.append(item.strip())
-
-    source = payload.get("source", "")
-    if not isinstance(source, str) or not source.strip():
-        return _build_error_response("source is required and must be a non-empty string")
-    source = source.strip()
-
-    language = payload.get("language", "en")
-    if not isinstance(language, str) or not language.strip():
-        return _build_error_response("language must be a non-empty string")
-    language = language.strip().lower()
-
-    results: List[Dict[str, Any]] = []
-    created_count = 0
-    already_exists_count = 0
-
-    for sentence_text in sentences:
-        existing = find_sentence_by_text(g.db, sentence_text, language_code=language)
-        if existing is not None:
-            results.append(
-                {
-                    "sentence_text": sentence_text,
-                    "sentence_id": existing.id,
-                    "guid": existing.guid,
-                    "status": "already_exists",
-                }
-            )
-            already_exists_count += 1
-            continue
-
-        sentence_row = add_sentence(g.db, source_filename=source)
-        add_sentence_translation(
-            g.db,
-            sentence=sentence_row,
-            language_code=language,
-            translation_text=sentence_text,
-            verified=False,
-        )
-        results.append(
-            {
-                "sentence_text": sentence_text,
-                "sentence_id": sentence_row.id,
-                "guid": sentence_row.guid,
-                "status": "created",
-            }
-        )
-        created_count += 1
-
-    return _build_success_response(
-        results,
-        {
-            "total": len(sentences),
-            "created": created_count,
-            "already_exists": already_exists_count,
-            "source": source,
-            "language": language,
-        },
-    )
-
-
-@bp.route("/v1/models")
-@mirrored_facade("/api/v1/models", "GET")
-def list_models() -> ResponseReturnValue:
-    """Search LLM models registered in the benchmarks database.
-
-    Query parameters:
-        - q: Optional. Case-insensitive substring search across codename, displayname,
-             model_path, and lmstudio_model_name fields.
-    """
-    bench_db = g.get("bench_db")
-    if bench_db is None:
-        return _build_error_response(
-            "Model registry not available (benchmarks database not configured)", 503
-        )
-
-    from benchmarks.datastore.common import Model
-
-    query = bench_db.query(Model).order_by(Model.codename)
-
-    search_term = request.args.get("q", "").strip()
-    if search_term:
-        pattern = f"%{search_term}%"
-        query = query.filter(
-            or_(
-                Model.codename.ilike(pattern),
-                Model.displayname.ilike(pattern),
-                Model.model_path.ilike(pattern),
-                Model.lmstudio_model_name.ilike(pattern),
-            )
-        )
-
-    models = query.all()
-
-    models_data = [
-        {
-            "codename": m.codename,
-            "displayname": m.displayname,
-            "model_path": m.model_path,
-            "model_type": m.model_type,
-            "lmstudio_model_name": m.lmstudio_model_name,
-            "launch_date": str(m.launch_date) if m.launch_date else None,
-            "license_name": m.license_name,
-        }
-        for m in models
-    ]
-
-    return _build_success_response(
-        models_data,
-        {"count": len(models_data), "search": search_term or None},
-    )
