--- src/barsukas/routes/api/v1_core.py
@@ -1,3 +1,0 @@
-"""Compatibility shim; routes live in lemma_routes.py."""
-
-from barsukas.routes.api.lemma_routes import *  # noqa: F401,F403
