--- src/barsukas/routes/api/v1_meta.py
@@ -1,3 +1,0 @@
-"""Compatibility shim; routes live in meta_routes.py."""
-
-from barsukas.routes.api.meta_routes import *  # noqa: F401,F403
