--- src/barsukas/routes/api/v1_sentences.py
@@ -1,3 +1,0 @@
-"""Compatibility shim; routes live in sentence_routes.py."""
-
-from barsukas.routes.api.sentence_routes import *  # noqa: F401,F403
